Updated project to work when direct navigating

+++ web/src/store/modules/projects.js
@@ -3,12 +3,15 @@ import axios from 'axios'
 const baseUrl = 'https://localhost:5001'
 
 const state = () => ({
-  all: []
+  all: [],
+  project: undefined
 })
 
 const getters = {
   find (state) {
-    return (id) => state.all.find(project => project.id === id)
+    return (id) => {
+      state.all.find(project => project.id === id)
+    }
   }
 }
 
@@ -21,8 +24,28 @@ const actions = {
       console.error(error)
     }
   },
+  async find ({ commit, state }, projectId) {
+    try {
+      if (!projectId) {
+        console.error('Empty ProjectId')
+        return
+      }
+      if (state.all.length) {
+        commit('setProject', state.all.find(project => project.id === projectId))
+      } else {
+        const response = await axios.get(baseUrl + '/projects/' + projectId)
+        commit('setProject', response.data)
+      }
+    } catch (error) {
+      console.error(error)
+    }
+  },
   async create ({ commit }, project) {
     try {
+      if (!project) {
+        console.error('Empty Project object')
+        return
+      }
       const response = await axios.post(baseUrl + '/projects', project)
       commit('addProject', response.data)
     } catch (error) {
@@ -31,6 +54,10 @@ const actions = {
   },
   async update ({ commit }, project) {
     try {
+      if (!project) {
+        console.error('Empty Project object')
+        return
+      }
       const response = await axios.patch(baseUrl + '/projects', project)
       commit('updateProject', response.data)
     } catch (error) {
@@ -39,6 +66,10 @@ const actions = {
   },
   async delete ({ commit }, projectId) {
     try {
+      if (!projectId) {
+        console.error('Empty ProjectId')
+        return
+      }
       await axios.delete(baseUrl + '/projects/' + projectId)
       commit('deleteProject', projectId)
     } catch (error) {
@@ -51,11 +82,15 @@ const mutations = {
   setProjects (state, projects) {
     state.all = projects
   },
+  setProject (state, project) {
+    state.project = project
+  },
   addProject (state, project) {
     state.all.push(project)
   },
   updateProject (state, project) {
-
+    const index = state.all.findIndex(p => p.id === project.id)
+    state.all[index] = project
   },
   deleteProject (state, projectId) {
     state.all = state.all.filter(project => project.id !== projectId)
